Add no-dict-subscript checker
